Am adaugat mai multe mesaje inregistrare refactorizare App separare admin/moderator/user/nelogat imagine default 90% gata admin Users(paginare,cautare,stergere)

+++ Project-tW/models/user.model.php
@@ -4,52 +4,72 @@ use LinkedIn\AccessToken;
 
 class UserModel
 {
-    private $db;
 
-    /**
-     * UserModel constructor.
-     * @param $db
-     */
-    public function __construct(PDO $db)
+    public function test()
     {
-        $this->db = $db;
+        $stmt = $this->db->prepare("INSERT INTO users(username) values('dasda')");
+        for ($i = 1; $i < 1000; $i++) {
+            $stmt->execute();
+        }
     }
 
-    public function getUsers()
+    public function getMaxPage(): int
     {
-        $stmt = $this->db->query("SELECT id,username,e_mail,last_name,first_name,linkedln_token ,github_token  FROM users");
-        $result = $stmt->fetchAll(PDO::FETCH_CLASS, "UserAdmin");
-        return $result;
+        $stmt = $this->db->prepare("SELECT count(id) from users where  id!=1");
+        $stmt->execute();
+        $lines = $stmt->fetch(PDO::FETCH_ASSOC)['count(id)'];
+        return $lines % Constants::MAX_PAGE > 0 ? ($lines / Constants::MAX_PAGE) + 1 : $lines / Constants::MAX_PAGE;
     }
 
-//    todo   search users
-    public function getUsersSearch(string $search)
+    public function getMaxPageSearch($search): int
     {
+        $stmt = $this->db->prepare("SELECT count(id) from users where (username like ? or e_mail like ? or last_name like ? or first_name like ?) and id!=1");
+        $stmt->execute([$search, $search, $search, $search]);
+        $lines = $stmt->fetch(PDO::FETCH_ASSOC)['count(id)'];
+        return $lines % Constants::MAX_PAGE > 0 ? ($lines / Constants::MAX_PAGE) + 1 : $lines / Constants::MAX_PAGE;
+    }
+
+    public function getUserGrade(string $user)
+    {
+        $stmt = $this->db->prepare("SELECT grade from users where username=?");
+        $stmt->execute([$user]);
+        return $stmt->fetch();
+    }
+
+    public function getUsers(int $page)
+    {
+        $page = ($page - 1) * Constants::MAX_PAGE;
+        $stmt = $this->db->prepare("SELECT id,username,e_mail,last_name,first_name,linkedln_token ,github_token  FROM users where id!=1 LIMIT ?,30");
+        $stmt->setFetchMode(PDO::FETCH_CLASS, "UserAdmin");
+        $stmt->execute([$page]);
+        return $stmt->fetchAll();
+    }
+
+    public function getUsersSearch(string $search, $page)
+    {
+        $search = '%' . $search . '%';
+
+        $page = ($page - 1) * Constants::MAX_PAGE;
         $stmt = $this->db->prepare("SELECT id,username,e_mail,last_name,first_name,linkedln_token ,github_token
-                                                FROM users where username like :usern");
-        $stmt->bindParam(":usern", $search);
-//        $stmt->bindParam(":email", $search);
-//        $stmt->bindParam(":lastn", $search);
-//        $stmt->bindParam(":firstn", $search);
-        $result = $stmt->fetch();
-        var_dump($result);
-//        if (!$result) {
-//            return array();
-//        }
-        return $result;
+                                                FROM users where (username like ? or e_mail like ? or last_name like ? or first_name like ?) and id!=1 limit ?,30");
+        $stmt->setFetchMode(PDO::FETCH_CLASS, "UserAdmin");
+        $stmt->execute([$search, $search, $search, $search, $page]);
+        $result = $stmt->fetchAll();
+        return $result == FALSE ? array() : $result;
     }
 
     public function deleteById($id)
     {
-        $stmt = $this->db->prepare("DELETE FROM users where id=:id");
-        $stmt->bindValue(":id", $id);
-        $stmt->execute();
+        if ($id != 1) {
+            $stmt = $this->db->prepare("DELETE FROM users where id=:id");
+            $stmt->bindValue(":id", $id);
+            $stmt->execute();
+        }
     }
 
 
     public function getTokenGithub(string $user)
     {
-
         $stmt = $this->db->prepare("SELECT * FROM users WHERE username=:user");
         $stmt->bindValue(":user", $user);
         $stmt->setFetchMode(PDO::FETCH_CLASS, "UserEntity");
@@ -93,10 +113,7 @@ class UserModel
         $stmt->setFetchMode(PDO::FETCH_NUM);
         $stmt->execute();
         $result = $stmt->fetch();
-        if ($result != NULL) {
-            return FALSE;
-        }
-        return TRUE;
+        return $result[0]!=NULL;
     }
 
     public function isConnectedWithLinkedln(string $user): bool
@@ -106,10 +123,7 @@ class UserModel
         $stmt->setFetchMode(PDO::FETCH_NUM);
         $stmt->execute();
         $result = $stmt->fetch();
-        if (!$result != NULL) {
-            return FALSE;
-        }
-        return TRUE;
+        return $result[0]!=NULL;
     }
 
     public function getLinkedlnToken(string $user): TokenLinkedln
@@ -175,7 +189,6 @@ class UserModel
             return UserEntity::default();
         }
         return $result;
-
     }
 
     public function saveUser(Register $user)
@@ -183,7 +196,7 @@ class UserModel
         $stmt = $this->db->prepare("INSERT INTO users(username, e_mail, password, first_name, last_name ) VALUES(?,?,?,?,?)");
         $stmt->bindValue(1, $user->getUsername());
         $stmt->bindValue(2, $user->getEmail());
-        $stmt->bindValue(3, password_hash($user->getPassword(),PASSWORD_DEFAULT));
+        $stmt->bindValue(3, password_hash($user->getPassword(), PASSWORD_DEFAULT));
         $stmt->bindValue(4, $user->getName());
         $stmt->bindValue(5, $user->getLastName());
         $stmt->execute();
@@ -196,5 +209,17 @@ class UserModel
         $stmt->execute();
         return $stmt->fetch() != FALSE;
     }
+
+    /**
+     * UserModel constructor.
+     * @param $db
+     */
+    public function __construct(PDO $db)
+    {
+        $this->db = $db;
+    }
+
+    private $db;
+
 
 }
